nodes2eventlog: Allow to specify offline detection threshold
Currently all nodes are detected immediatelly as being offline. This can be
confusing for nodes which only have a short interruption in their
connectivity. Most of the time only nodes with long-lasting problems need
manual interventions.

This threshold only affects the state database and already existing
entries in the eventlog will stay there.