nodes2eventlog: Keep a log of offline entries
The eventlog only contains the events about nodes which went offline and
the state information only contain the nodes which are not yet offline for
too long. The graveyard.pickle contains the list of nodes which are offline
and were offline when they were removed from the states.pickle.

The helper graveyard2rst.py can be used to convert the information to a
readable list.